+++ packages/stg-evm-v2/devtools/config/mainnet/01/asset.usdc.config.ts
@@ -29,6 +29,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const flarePoint = getAssetPoint(EndpointId.FLARE_V2_MAINNET)
     const gravityPoint = getAssetPoint(EndpointId.GRAVITY_V2_MAINNET)
     const iotaPoint = getAssetPoint(EndpointId.IOTA_V2_MAINNET)
+    const islanderPoint = getAssetPoint(EndpointId.ISLANDER_V2_MAINNET)
     const klaytnPoint = getAssetPoint(EndpointId.KLAYTN_V2_MAINNET)
     const lightlinkPoint = getAssetPoint(EndpointId.LIGHTLINK_V2_MAINNET)
     const mantlePoint = getAssetPoint(EndpointId.MANTLE_V2_MAINNET)
@@ -39,6 +40,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const rariblePoint = getAssetPoint(EndpointId.RARIBLE_V2_MAINNET)
     const scrollPoint = getAssetPoint(EndpointId.SCROLL_V2_MAINNET)
     const seiPoint = getAssetPoint(EndpointId.SEI_V2_MAINNET)
+    const superpositionPoint = getAssetPoint(EndpointId.SUPERPOSITION_V2_MAINNET)
     const taikoPoint = getAssetPoint(EndpointId.TAIKO_V2_MAINNET)
     const xchainPoint = getAssetPoint(EndpointId.XCHAIN_V2_MAINNET)
 
@@ -54,6 +56,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const flareContract = await getAssetNode(flarePoint)
     const gravityContract = await getAssetNode(gravityPoint)
     const iotaContract = await getAssetNode(iotaPoint)
+    const islanderContract = await getAssetNode(islanderPoint)
     const klaytnContract = await getAssetNode(klaytnPoint)
     const lightlinkContract = await getAssetNode(lightlinkPoint)
     const mantleContract = await getAssetNode(mantlePoint)
@@ -64,6 +67,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
     const raribleContract = await getAssetNode(rariblePoint)
     const scrollContract = await getAssetNode(scrollPoint)
     const seiContract = await getAssetNode(seiPoint)
+    const superpositionContract = await getAssetNode(superpositionPoint)
     const taikoContract = await getAssetNode(taikoPoint)
     const xchainContract = await getAssetNode(xchainPoint)
 
@@ -80,6 +84,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             flareContract,
             gravityContract,
             iotaContract,
+            islanderContract,
             klaytnContract,
             lightlinkContract,
             mantleContract,
@@ -90,6 +95,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             raribleContract,
             scrollContract,
             seiContract,
+            superpositionContract,
             taikoContract,
             xchainContract,
         ],
@@ -105,6 +111,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             flarePoint,
             gravityPoint,
             iotaPoint,
+            islanderPoint,
             klaytnPoint,
             lightlinkPoint,
             mantlePoint,
@@ -115,6 +122,7 @@ export default async (): Promise<OmniGraphHardhat<AssetNodeConfig, AssetEdgeConf
             rariblePoint,
             scrollPoint,
             seiPoint,
+            superpositionPoint,
             taikoPoint,
             xchainPoint,
         ]),